Jog wheel moves time back and forward but pause doesn't take the first time you use wheel

+++ Frameworks/MoonKit/Moonkit/Objects/ScrollWheel.swift
@@ -193,7 +193,7 @@ import UIKit
   }
 
   private var thetaOffset: CGFloat = 0
-  public var revolutions: CGFloat { return (theta - thetaOffset) / (π * 2) }
+  public var revolutions: Float { return Float((theta - thetaOffset) / (π * 2)) }
 
   @IBInspectable public var theta: CGFloat = 0 {
     didSet {
